added repository structure

+++ src/main/java/com/example/dao/GenericDAO.java
@@ -6,8 +6,9 @@ import java.util.List;
  * Created by yshpyluk on 4/5/17.
  */
 public interface GenericDAO<T> {
+
 	List<T> getAll();
-	T get(int i);
+	T get(Long i);
 	void add(T entity);
 	void remove(T entity);
 	T update(T entity);
